- Write all unit tests for GetCurrentWeatherForecastUseCase, - Emit CurrentWeatherForecastUiState from usecase
