+++ forecast/src/test/java/com/edwardmuturi/forecast/data/FakerForecastRepository.kt
@@ -1,0 +1,60 @@
+/*
+ * Copyright 2024
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining
+ * a copy of this software and associated documentation files (the
+ * "Software"), to deal in the Software without restriction, including
+ * without limitation the rights to use, copy, modify, merge, publish,
+ * distribute, sublicense, and/or sell copies of the Software, and to
+ * permit persons to whom the Software is furnished to do so, subject to
+ * the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be
+ * included in all copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
+ * EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
+ * MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
+ * NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE
+ * LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION
+ * OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
+ * WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+ */
+package com.edwardmuturi.forecast.data
+
+import com.edwardmuturi.forecast.data.remote.dto.FetchCurrentWeatherDataDto
+import com.edwardmuturi.forecast.data.remote.dto.FetchFiveDayForecastDto
+import com.edwardmuturi.forecast.data.sampledata.currentForecast
+import com.edwardmuturi.forecast.domain.repository.ForeCastRepository
+import kotlinx.coroutines.flow.Flow
+import kotlinx.coroutines.flow.flow
+
+class FakerForecastRepository : ForeCastRepository {
+    private var result: Result<FetchCurrentWeatherDataDto?> = Result.success(currentForecast)
+
+    fun setSuccessfullResult() {
+        result = Result.success(currentForecast)
+    }
+
+    fun setSuccessfullNullResult() {
+        result = Result.success(null)
+    }
+
+    fun setErrorResult() {
+        result = Result.failure(Exception("Failed to parse response body"))
+    }
+
+    override fun getCurrentDayForecast(
+        lat: String,
+        lon: String
+    ): Flow<Result<FetchCurrentWeatherDataDto?>> {
+        return flow { emit(result) }
+    }
+
+    override fun getFiveDayForecast(
+        lat: String,
+        lon: String
+    ): Flow<Result<FetchFiveDayForecastDto?>> {
+        TODO("Not yet implemented")
+    }
+}
