+++ forecast/src/test/java/com/edwardmuturi/forecast/data/sampledata/SampleData.kt
@@ -20,28 +20,61 @@
  * OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
  * WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
-package com.edwardmuturi.forecast.presentation
+package com.edwardmuturi.forecast.data.sampledata
 
-import androidx.lifecycle.ViewModel
-import androidx.lifecycle.viewModelScope
-import com.edwardmuturi.forecast.domain.usecase.GetCurrentWeatherForecastUseCase
-import dagger.hilt.android.lifecycle.HiltViewModel
-import javax.inject.Inject
-import kotlinx.coroutines.flow.collectLatest
-import kotlinx.coroutines.launch
-import timber.log.Timber
+import com.edwardmuturi.forecast.data.remote.dto.FetchCurrentWeatherDataDto
+import com.google.gson.Gson
 
-@HiltViewModel
-class ForecastViewModel @Inject constructor(
-    private val getCurrentWeatherForecastUseCase: GetCurrentWeatherForecastUseCase
-) :
-    ViewModel() {
-    fun loadCurrentDayForecast() {
-        viewModelScope.launch {
-            getCurrentWeatherForecastUseCase(lat = "44.34", lon = "10.99")
-                .collectLatest {
-                    Timber.e("Result it $it")
-                }
-        }
+val currentForecastString = """
+{
+  "coord": {
+    "lon": 10.99,
+    "lat": 44.34
+  },
+  "weather": [
+    {
+      "id": 501,
+      "main": "Rain",
+      "description": "moderate rain",
+      "icon": "10d"
     }
-}
+  ],
+  "base": "stations",
+  "main": {
+    "temp": 298.48,
+    "feels_like": 298.74,
+    "temp_min": 297.56,
+    "temp_max": 300.05,
+    "pressure": 1015,
+    "humidity": 64,
+    "sea_level": 1015,
+    "grnd_level": 933
+  },
+  "visibility": 10000,
+  "wind": {
+    "speed": 0.62,
+    "deg": 349,
+    "gust": 1.18
+  },
+  "rain": {
+    "1h": 3.16
+  },
+  "clouds": {
+    "all": 100
+  },
+  "dt": 1661870592,
+  "sys": {
+    "type": 2,
+    "id": 2075663,
+    "country": "IT",
+    "sunrise": 1661834187,
+    "sunset": 1661882248
+  },
+  "timezone": 7200,
+  "id": 3163858,
+  "name": "Zocca",
+  "cod": 200
+}                        
+""".trimIndent()
+
+val currentForecast = Gson().fromJson(currentForecastString, FetchCurrentWeatherDataDto::class.java)
